nsfplay fixed frequency rounding for DMC

+++ trunk/xgm/devices/Sound/nes_dmc.cpp
@@ -375,7 +375,7 @@ namespace xgm
 
   void NES_DMC::SetClock (double c)
   {
-    clock = (UINT32)(c);
+    clock = c;
   }
 
   void NES_DMC::SetRate (double r)
